GUI/requirements.txt
asgiref==3.7.2
nibabel==5.2.1
SimpleITK==2.3.1
opencv-python==4.9.0
matplotlib==3.8.3
asn1==2.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
customtkinter==5.2.2
darkdetect==0.8.0
distlib==0.3.7
easygui==0.98.3
enum-compat==0.0.3
filelock==3.12.4
fsspec==2024.2.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pillow==10.2.0
platformdirs==3.10.0
pycryptodome==3.18.0
pyimg4==0.8
sympy==1.12
urllib3==2.2.1
virtualenv==20.24.5
